+++ src/Exceptions/GeometryExceptionInterface.php
@@ -1,0 +1,11 @@
+<?php
+namespace samizdam\Geometry\Exceptions;
+
+/**
+ * 
+ * @author samizdam
+ *
+ */
+interface GeometryExceptionInterface
+{
+}