+++ src/Exceptions/InvalidArgumentException.php
@@ -1,0 +1,11 @@
+<?php
+namespace samizdam\Geometry\Exceptions;
+
+/**
+ *
+ * @author samizdam
+ *        
+ */
+class InvalidArgumentException extends \InvalidArgumentException implements GeometryExceptionInterface
+{
+}