+++ src/Exceptions/OutOfBoundsException.php
@@ -1,0 +1,11 @@
+<?php
+namespace samizdam\Geometry\Exceptions;
+
+/**
+ *
+ * @author samizdam
+ *        
+ */
+class OutOfBoundsException extends \OutOfBoundsException implements GeometryExceptionInterface
+{
+}