+++ src/Plane/Angle/AngleFactory.php
@@ -1,0 +1,18 @@
+<?php
+namespace samizdam\Geometry\Plane\Angle;
+
+/**
+ *
+ * @author samizdam
+ *        
+ */
+class AngleFactory implements AngleFactoryInterface
+{
+
+    private $defaultUnit;
+
+    public function getDefaultUnit()
+    {
+        return $this->defaultUnit ?  : $this->defaultUnit = new AngleSizeUnit(AngleSizeUnit::DEFAULT_TYPE);
+    }
+}