+++ src/Plane/Angle/AngleFactoryInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace samizdam\Geometry\Plane\Angle;
+
+interface AngleFactoryInterface
+{
+    public function getDefaultUnit();
+}