+++ src/Plane/Angle/AngleSizeCalculator.php
@@ -2,16 +2,30 @@
 namespace samizdam\Geometry\Plane\Angle;
 
 use samizdam\Geometry\Plane\ConstantsAwareTrait;
+use samizdam\Geometry\Plane\FactoriesCollectionAwareTrait;
 
 /**
  *
  * @author samizdam
  *        
  */
-class AngleSizeCalculationStrategy implements AngleSizeCalculationStrategyInterface
+class AngleSizeCalculator implements AngleSizeCalculatorInterface
 {
     
     use ConstantsAwareTrait;
+    use FactoriesCollectionAwareTrait;
+
+    private $angleSizeUnit;
+
+    public function setAngleSizeUnit(AngleSizeUnitsEnum $unit)
+    {
+        $this->angleSizeUnit = $unit;
+    }
+
+    public function getAngleSizeUnit()
+    {
+        return $this->angleSizeUnit ?  : $this->angleSizeUnit = $this->getFactoriesCollection()->getAngleFactory()->getDefaultUnit();
+    }
 
     public function getAngleSize(AngleInterface $angle)
     {