+++ src/Plane/Angle/AngleSizeCalculatorInterface.php
@@ -1,8 +1,10 @@
 <?php
 namespace samizdam\Geometry\Plane\Angle;
 
-interface AngleSizeCalculationStrategyInterface
+interface AngleSizeCalculatorInterface
 {
+    
+    public function setAngleSizeUnit(AngleSizeUnitsEnum $unit);
     
     public function getAngleSize(AngleInterface $angle);
 }