+++ src/Plane/FactoriesCollection.php
@@ -5,6 +5,9 @@ use samizdam\Geometry\Plane\Polygons\PolygonFactoryInterface;
 use samizdam\Geometry\Plane\Polygons\PolygonFactory;
 use samizdam\Geometry\Plane\Lines\LineFactoryInterface;
 use samizdam\Geometry\Plane\Lines\LineFactory;
+use samizdam\Geometry\Plane\Angle\AngleFactoryInterface;
+use samizdam\Geometry\Plane\Angle\AngleFactory;
+use samizdam\Geometry\Exceptions as Exceptions;
 
 /**
  *
@@ -17,6 +20,7 @@ class FactoriesCollection implements FactoriesCollectionInterface
     private $objectPool;
 
     private $classMap = [
+        AngleFactoryInterface::class => AngleFactory::class,
         PolygonFactoryInterface::class => PolygonFactory::class,
         LineFactoryInterface::class => LineFactory::class
     ];
@@ -31,7 +35,7 @@ class FactoriesCollection implements FactoriesCollectionInterface
         if (! $this->objectPool->offsetExists($interface) && isset($this->classMap[$interface])) {
             $this->setFactory($interface, $this->classMap[$interface]);
         } else {
-            throw new \OutOfBoundsException("Unknow factory interface {$interface}");
+            throw new Exceptions\OutOfBoundsException("Unknow factory interface {$interface}");
         }
         return $this->objectPool->offsetGet($interface);
     }
@@ -42,8 +46,13 @@ class FactoriesCollection implements FactoriesCollectionInterface
         if ($factoryInstance instanceof $interface) {
             $this->objectPool->offsetSet($interface, $factoryInstance);
         } else {
-            throw new \InvalidArgumentException("{$factoryClassName} must implement {$interface}.");
+            throw new Exceptions\InvalidArgumentException("{$factoryClassName} must implement {$interface}.");
         }
+    }
+
+    public function getAngleFactory()
+    {
+        return $this->getFactory(AngleFactoryInterface::class);
     }
 
     public function getPolygonFactory()