+++ src/Plane/Lines/RayInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace samizdam\Geometry\Plane\Lines;
+
+interface RayInterface
+{
+    public function getFirstPoint();
+}