+++ src/Plane/PlaneGeometry.php
@@ -21,16 +21,26 @@ class PlaneGeometry implements PlaneGeometryInterface
 
     private $factoriesCollection;
 
-    public function __construct(FactoriesCollectionInterface $factoryCollection = null, CalculationStrategiesCollectionInterface $calculationStrategiesCollection = null, Constants $constants = null)
+    public function __construct(FactoriesCollectionInterface $factoryCollection = null, ComposeCalculatorInterface $calculationStrategiesCollection = null, Constants $constants = null)
     {
         $this->constants = $constants ?  : new Constants();
-        $this->calculationStrategiesCollection = $calculationStrategiesCollection ?  : new CalculationStrategiesCollection();
+        $this->calculationStrategiesCollection = $calculationStrategiesCollection ?  : new ComposeCalculator();
         $this->factoriesCollection = $factoryCollection ?  : new FactoriesCollection();
     }
 
     public function createPolygonByPoints(array $points)
     {
         return $this->factoriesCollection->getPolygonFactory()->createPolygonByPoints($points);
+    }
+
+    public function createLine(PointInterface $A, PointInterface $B)
+    {
+        return $this->factoriesCollection->getLineFactory()->createLine($A, $B);
+    }
+
+    public function createRay(PointInterface $pointA, PointInterface $pointB)
+    {
+        return $this->factoriesCollection->getLineFactory()->createRay($pointA, $pointB);
     }
 
     public function createLineSegment(PointInterface $A, PointInterface $B)