+++ src/Plane/PlaneGeometryInterface.php
@@ -25,7 +25,23 @@ interface PlaneGeometryInterface
      *
      * @param PointInterface $A
      * @param PointInterface $B
-     * @return Lines\LineSegment
+     * @return Lines\LineSegmentInterface
      */
     public function createLineSegment(PointInterface $A, PointInterface $B);
+
+    /**
+     *
+     * @param PointInterface $A
+     * @param PointInterface $B
+     * @return Lines\LineInteface
+     */
+    public function createLine(PointInterface $A, PointInterface $B);
+
+    /**
+     *
+     * @param PointInterface $pointA
+     * @param PointInterface $pointB
+     * @return Lines\RayInterface
+     */
+    public function createRay(PointInterface $pointA, PointInterface $pointB);
 }