+++ src/Plane/Polygons/PolygonFactory.php
@@ -1,9 +1,16 @@
 <?php
 namespace samizdam\Geometry\Plane\Polygons;
 
+/**
+ *
+ * @author samizdam
+ *        
+ */
 class PolygonFactory implements PolygonFactoryInterface
 {
 
     public function createPolygonByPoints(array $points)
-    {}
+    {
+        return AbstractPolygon::createByPoints($points);
+    }
 }