+++ tests/Plane/CalculatorAwareTraitTest.php
@@ -1,0 +1,18 @@
+<?php
+namespace samizdam\Geometry\Plane;
+
+use samizdam\Geometry\GeometryUnitTestCase;
+
+/**
+ * 
+ * @author samizdam
+ *
+ */
+class CalculatorAwareTraitTest extends GeometryUnitTestCase
+{
+    public function testSetter()
+    {
+        $calc = $this->getMockForTrait(CalculatorAwareTrait::class);
+        $calc->setComposeCalculator($this->getMock(ComposeCalculatorInterface::class));
+    }
+}