+++ CHANGELOG.md
@@ -15,8 +15,21 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Normalize naming of arguments, fields and methods. To fix in some guideline.  
 - Extract LineSegment length calculation. 
 
+## 0.3.0 - 2015-09-13
+### Change 
+- Rename calcalation classes and methods. KISS =)
+- Rename Ray::getPointA() to getFirstPoint(). 
+
+### Added
+- AngleSizeUnit classes. 
+- AngleFactory classes. 
+- LineInterface.
+- RayInterface.
+- Exceptions namespace. 
+
 ### Fixed
 - Up testing coverage to 100%.  
+
 
 ## 0.2.0 - 2015-09-11
 ### Added